add uploader

+++ config/config.go
@@ -584,7 +584,7 @@ func (c *TelegramConf) Validate() (errs []error) {
 
 // SaasConf is stride config
 type SaasConf struct {
-	GroupID int    `json:"-"`
+	GroupID int64  `json:"-"`
 	Token   string `json:"-"`
 	URL     string `json:"-"`
 }
